+++ mm-video/vidc/vdec/inc/qtypes.h
@@ -64,11 +64,11 @@ extern "C" {
 
 typedef  unsigned char      boolean;     /* Boolean value type. */
 
-typedef  unsigned long int  uint32;      /* Unsigned 32 bit value */
+typedef  unsigned int  uint32;      /* Unsigned 32 bit value */
 typedef  unsigned short     uint16;      /* Unsigned 16 bit value */
 typedef  unsigned char      uint8;       /* Unsigned 8  bit value */
 
-typedef  signed long int    int32;       /* Signed 32 bit value */
+typedef  int    int32;       /* Signed 32 bit value */
 typedef  signed short       int16;       /* Signed 16 bit value */
 typedef  signed char        int8;        /* Signed 8  bit value */
